TOH.c
Tower of Hanoi is a mathematical puzzle where we have three rods and n disks. The objective of the puzzle is to move the entire stack to another rod, obeying the following simple rules:
i. Only one disk can be moved at a time.
ii. Each move consists of taking the upper disk from one of the stacks and placing it on top of another stack i.e. a disk can only be moved if it is the uppermost disk on a stack.
iii. No disk may be placed on top of a smaller disk
iv. You can choose to use the function move (4, 1, 3, 2), where 4 represents the number of disks. 1 represents disks on source shaft, 3 represents the destination shaft which holds the disks after the move and finally 2 represents the intermediate support shaft – temporary storage. 
Write a C program to simulate the given problem and: Perform the algorithmic complexity analysis for the solution you propose.