simulation.c
Let us consider a small but busy airport with only one run-way (shown in figure). In each time unit, one plane can land or one plane can take off, but not both. Planes arrive ready to land or to take off at random times, so at any given unit of time, the runway
may be idle or a plan may be landing or taking off, and there may be several planes waiting
either to land or take off. We therefore need two queues, called landing and takeoff, to hold
these planes. It is better to keep a plane waiting on the ground than in the air, so a small
airport allows a plane to take off only if there are no planes waiting to land. Hence, after
receiving requests from new planes to land or take off, our simulation will first service the
head of the queue of planes waiting to land, and only if the landing queue is empty will it
allow a plane to take off. We shall wish to run the simulation through many units of time,
and therefore, we embed the main action of the program in a loop that runs for cur-time
(denoting current time) from 1 to a variable end-time.
Simulate the given scenario using and write the output for different inputs.