Dijkstra.c
The Dijkstra’s algorithm is an algorithm that gives the shortest path between two given vertices of a graph. In this problem we are given a directed graph with each edge having a non-negative weight. Thus, a solution requires a path of many other that
costs least. We can think of the problem as like this: think graph G as a map of the airline routes, each node of the graph as the cities and the weights on each edge as the cost of flying
from one city to another city. The solution we have to find a routing from a city v to city w
such that the total cost is minimum. Write a C program to simulate the given problem. That is
find the shortest path between node A and node F in the given graph.